+++ actix-tls/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `rustls-0_23` crate feature which excludes any root certificate methods or re-exports.
+- Minimum supported Rust version (MSRV) is now 1.70.
 
 ## 3.3.0